refactor: move DropDownOption class out of ElfRelocatableObjectExporter
